# ai-specs/specs/api-spec.yml
openapi: 3.0.3
info:
  title: Proyecto Musical API
  version: 0.1.0
  description: |
    Contrato de los endpoints HTTP expuestos por la aplicacion Next.js 15.
    La mayoria de mutaciones del MVP se hacen via server actions y no aparecen
    aqui; este documento cubre endpoints REST accesibles desde cliente.
paths:
...
  /api/auth/register:
    post:
      summary: Registro de usuario
      description: Crea un usuario con email + password. Devuelve el usuario creado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                name: { type: string, maxLength: 80 }
                email: { type: string, format: email, maxLength: 180 }
                password: { type: string, minLength: 8, maxLength: 128 }
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id: { type: string }
                      email: { type: string }
                      name: { type: string, nullable: true }
        '400': { description: Datos invalidos }
        '409': { description: Email ya registrado }
...